Fix check for forwarded_ports
We can't check if there are forwarded ports defined because there will
always be the ssh forwarded port defined by vagrant. Instead, we need
to check if there are no networks defined. That means the host networking
setup is commented out in the vagrantfile. By checking the networks, we
require that port forwarding (for port 80) and host networking not be
used at the same time.